refactored_v1
