--- src/main/java/com/gsoft/projectManager/appuser/AppUserRole.java
@@ -1,7 +1,0 @@
-package com.gsoft.projectManager.appuser;
-
-public enum AppUserRole {
-    ADMIN,
-    MNGR,
-    DEV
-}
