+++ src/main/java/com/gsoft/projectManager/model/AppUser.java
@@ -1,4 +1,4 @@
-package com.gsoft.projectManager.appuser;
+package com.gsoft.projectManager.model;
 
 
 import java.util.List;
