+++ src/main/java/com/gsoft/projectManager/model/UserPrincipal.java
@@ -1,4 +1,4 @@
-package com.gsoft.projectManager.appuser;
+package com.gsoft.projectManager.model;
 
 import java.util.ArrayList;
 import java.util.Collection;
