+++ src/main/java/com/gsoft/projectManager/repository/RoleRepository.java
@@ -1,7 +1,9 @@
-package com.gsoft.projectManager.appuser;
+package com.gsoft.projectManager.repository;
 
 import java.util.Optional;
 
+import com.gsoft.projectManager.model.Role;
+import com.gsoft.projectManager.model.Rolename;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
